requirements.txt
aiogram==3.0.0
aiohttp==3.8.5
pandas==2.1.0
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.0
python-telegram-bot==20.5
tabulate==0.9.0
